requirements.txt
streamlit==1.31.0
durable-rules==2.0.28
pytest==7.4.3